server/requirements.txt
rembg==2.0.67
pillow==10.4.0
